requirements.txt
aiogram==3.13.1
SQLAlchemy==2.0.35
aiosqlite==0.20.0
APScheduler==3.10.4
pytz==2024.1
python-dotenv==1.0.1
pytest==8.3.3
pytest-asyncio==0.24.0
